+++ web/src/components/modules/site/AccountEditDialog.tsx
@@ -118,6 +118,8 @@ function credentialOptions(platform: SitePlatform) {
         case SitePlatform.Sub2API:
             return [SiteCredentialType.AccessToken, SiteCredentialType.APIKey];
         case SitePlatform.API:
+        case SitePlatform.SiliconFlow:
+        case SitePlatform.Other:
             return [SiteCredentialType.AccessToken, SiteCredentialType.APIKey];
         default:
             return [